refactor: request, response 패키지 수준 증가
