+++ src/main/java/roomescape/controller/ReservationController.java
@@ -7,13 +7,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.request.ReservationRequest;
 import roomescape.annotation.AuthenticationPrincipal;
-import roomescape.controller.request.ReservationRequest;
-import roomescape.controller.response.MemberReservationResponse;
-import roomescape.controller.response.ReservationResponse;
 import roomescape.model.Member;
 import roomescape.model.Payment;
 import roomescape.model.Reservation;
+import roomescape.response.MemberReservationResponse;
+import roomescape.response.ReservationResponse;
 import roomescape.service.PaymentService;
 import roomescape.service.ReservationService;
 import roomescape.service.ReservationWaitingService;
