+++ src/main/java/roomescape/controller/ThemeController.java
@@ -1,10 +1,15 @@
 package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import roomescape.controller.request.ThemeRequest;
-import roomescape.controller.response.ReservationThemeResponse;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.request.ThemeRequest;
 import roomescape.model.Theme;
+import roomescape.response.ReservationThemeResponse;
 import roomescape.service.ThemeService;
 
 import java.net.URI;
