+++ src/main/java/roomescape/controller/WaitingController.java
@@ -1,12 +1,17 @@
 package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.request.WaitingRequest;
 import roomescape.annotation.AuthenticationPrincipal;
-import roomescape.controller.request.WaitingRequest;
-import roomescape.controller.response.WaitingResponse;
 import roomescape.model.Member;
 import roomescape.model.Waiting;
+import roomescape.response.WaitingResponse;
 import roomescape.service.WaitingService;
 
 import java.net.URI;
@@ -23,7 +28,7 @@ public class WaitingController {
 
     @PostMapping("/waiting")
     public ResponseEntity<WaitingResponse> createWaiting(@RequestBody WaitingRequest request,
-                                                 @AuthenticationPrincipal Member member) {
+                                                         @AuthenticationPrincipal Member member) {
         Waiting waiting = waitingService.addWaiting(request, member);
         WaitingResponse response = new WaitingResponse(waiting);
         return ResponseEntity.created(URI.create("/waiting/" + waiting.getId())).body(response);
