+++ src/main/java/roomescape/request/ReservationRequest.java
@@ -1,17 +1,20 @@
-package roomescape.controller.request;
+package roomescape.request;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.annotation.Nonnull;
 import jakarta.validation.constraints.Positive;
-import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDate;
 
-public record WaitingRequest(
+public record ReservationRequest(
         @Nonnull
-        @DateTimeFormat
+        @JsonFormat(pattern = "yyyy-MM-dd")
         LocalDate date,
         @Positive(message = "[ERROR] timeId의 값이 1보다 작을 수 없습니다.")
         long timeId,
         @Positive(message = "[ERROR] themeId의 값이 1보다 작을 수 없습니다.")
-        long themeId) {
+        long themeId,
+        String orderId,
+        String paymentKey,
+        long amount) {
 }
