+++ src/main/java/roomescape/service/ReservationService.java
@@ -2,8 +2,8 @@ package roomescape.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.controller.request.AdminReservationRequest;
-import roomescape.controller.request.ReservationRequest;
+import roomescape.request.AdminReservationRequest;
+import roomescape.request.ReservationRequest;
 import roomescape.exception.DuplicatedException;
 import roomescape.exception.NotFoundException;
 import roomescape.model.Member;
