+++ src/main/java/roomescape/service/ThemeService.java
@@ -2,7 +2,7 @@ package roomescape.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.controller.request.ThemeRequest;
+import roomescape.request.ThemeRequest;
 import roomescape.model.Theme;
 import roomescape.repository.ThemeRepository;
