--- src/test/java/roomescape/RoomescapeApplicationTest.java
@@ -1,12 +1,0 @@
-package roomescape;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-class RoomescapeApplicationTest {
-
-    @Test
-    void contextLoads() {
-    }
-}
