+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.BaseTest;
-import roomescape.controller.request.AdminReservationRequest;
-import roomescape.controller.request.ReservationRequest;
+import roomescape.request.AdminReservationRequest;
+import roomescape.request.ReservationRequest;
 import roomescape.exception.BadRequestException;
 import roomescape.exception.DuplicatedException;
 import roomescape.exception.NotFoundException;
